feat: Integrate Hive for local caching and update theme handling
This commit introduces Hive for local data caching, replacing the previous `Map`-based caching in `CacheService`. It also refactors theme handling in the main application widget.

**CacheService Changes (`lib/data/services/cache_service.dart`):**
- Integrated Hive for storing `Product` and `Rating` objects.
- Added `ProductAdapter` and `RatingAdapter` and ensured they are registered before opening Hive boxes.
- `_productBox` now stores `Product` objects directly instead of `Map`.
- `_categoryBox` now stores `List<String>` for categories.
- `cacheProducts`:
    - Stores `Product` objects in `_productBox`.
    - Stores product IDs as a comma-separated string in `_metaBox` under 'all_products'.
- `getCachedProducts`:
    - Retrieves product IDs from `_metaBox`.
    - Fetches `Product` objects directly from `_productBox` using the retrieved IDs.
- `getCachedProduct`: Fetches a `Product` directly from `_productBox` by ID.
- `cacheProduct`: Stores a `Product` object directly in `_productBox`.
- `getCachedCategories`: Retrieves `List<String>` directly from `_categoryBox`.

**Model Changes:**
- **`Product` (`lib/data/models/product.dart`, `lib/data/models/product.g.dart`):**
    - File deleted. (Presumably moved or replaced by a Hive-compatible version, though the new version is not in the diff).
- **`Rating` (`lib/data/models/rating.dart`, `lib/data/models/rating.g.dart`):**
    - File deleted. (Presumably moved or replaced by a Hive-compatible version).

**UI & App Logic Changes:**
- **`App` Widget (`lib/app/view/app.dart`):**
    - Wrapped `MaterialApp` with a `BlocBuilder<ThemeBloc, ThemeState>` to dynamically update the theme based on `state.isDarkMode`.
    - `ThemeData.dark()` or `ThemeData.light()` is applied accordingly.
- **`ProductDetailsPage` (`lib/ui/product_details/view/product_details_page.dart`):**
    - Updated price formatting to `state.product.price.toDouble().formatPrice`.
    - Updated rating display to use `state.product.rating.rate.toDouble()` and `state.product.rating.count.toInt()`.
- **`ProductCard` (`lib/ui/products/view/product_card.dart`):**
    - Cast `product.id` to `toInt()` when navigating to `ProductDetailsPage`.

**Repository & Service Updates:**
- **`ProductRepository` (`lib/data/repositories/product_repository.dart`):**
    - Added debug print statement for fetched products.
- **`ApiService` (`lib/data/services/api_service.dart`):**
    - Added debug print statements for the JSON list received from the API and the parsed `Product` objects.
    - Added error logging in `getProducts`.

**Dependency Changes (`pubspec.yaml`, `pubspec.lock`):**
- Added `hive_generator: ^2.0.1`.
- Added `json_annotation: ^4.9.0`.
- Added `json_serializable: ^6.9.0`.
- Added `build_runner: ^2.5.4` to dev_dependencies.
- Updated various transitive dependencies in `pubspec.lock`.

+++ lib/ui/product_details/view/product_details_page.dart
@@ -53,7 +53,7 @@ class ProductDetailsView extends StatelessWidget {
                     ),
                     const SizedBox(width: 16),
                     Text(
-                      state.product.price.formatPrice,
+                      state.product.price.toDouble().formatPrice,
                       style: Theme.of(context).textTheme.headlineMedium
                           ?.copyWith(
                             color: Colors.green,
@@ -88,8 +88,8 @@ class ProductDetailsView extends StatelessWidget {
                     ),
                     const Spacer(),
                     RatingWidget(
-                      rating: state.product.rating.rate,
-                      count: state.product.rating.count,
+                      rating: state.product.rating.rate.toDouble(),
+                      count: state.product.rating.count.toInt(),
                     ),
                   ],
                 ),
